Improve const-correctness by switching to ap_getword_nc
A couple of places, we use const char *s for dynamically-allocated
strings. This is incorrect, but made handling the second argument of
ap_getword much easier (its type const char **, which is a bit of a
swine).

We now use ap_getword_nc (which takes char ** instead), so can declare
the affected pointers as char * instead, which is more correct, and
requires less casting.

signed-off-by: Matthew Vernon <mcv21@cam.ac.uk>