Add support for Raven protocol version 3
The difference from protocol version 2 is the optional "ptags" flag
that the WLS may return. Currently, the only value that has meaning is
"current", meaning that the user being authenticated is a current
unversity person.

So we have a new required_ptags flag, which defaults to PTAGS_CURRENT
(i.e. require the "current" ptags to be set) - this is equivalent to
the v2 behaviour, as a v2 server will not authenticate users that the
v3 server would not give a "current" ptag to. Admins can set
AARequiredPtags to None (or just say "AARequiredPtags\n") to accept
users without the "current" ptag.

The patch does a number of things:

i) provide a safe atoui (simple parsing of a single unsigned int)
function

ii) handle the AARequiredPtags configuration command, check for ptags
in the webauth cookie and WLS responses if the server returns a V3 or
higher command.

iii) check that if we're using protocol version >=3 and the admin has
specified any ptags that the WLS response/cookie has the relevant
ptags

iv) in cases where authentication fails because of ptags mismatch,
pass a 403 (forbidden) error to the user; internally, we set code 601
in the session cookie (rather than the 600 used for other errors,
which end up as "bad request" for the user)

TODO: documentation!