Make dump_config work in multiple contexts
dump_config relied on a request_rec being available, thus using
apr_log_rerror. this is good style, but if we want to inspect the
merging process, that request_rec isn't available, we only have a pool
(so use apr_log_perror instead). These changes enable us to do so.