Correct UNSET values
UNSET was previously UINT_MAX, which doesn't do what you wanted, since
it has every bit set - what you wanted was a value with a bit set that
isn't used in any of the set values. Go with 1<<15 (32768) as that
should be fairly distinctive in the logs, and is guaranteed by C99 to
fit in an unsigned int.