Log at WARNING not DEBUG for config merging
This is a bit antisocial (apache will emit warnings when (re-)started,
for instance), but otherwise ap_log_perror's output gets lost
regardless of what LogLevel is specified in configuration files (see
Debian bug 714613 for more details of this).

We don't want to ship with this behaviour, but it's useful for testing.