Create AACookieForceSecure config option
There are times when users may wish to set the "secure" attribute of
the session cookie even when mod_ucam_webauth is running under http
(e.g. they are running behind a proxy that is doing the https). In
this case, if they turn AACookieForceSecure on, then that is what will
happen. Doing this if the browser is talking http will cause it not to
return the cookie.