Set AAPRINCIPAL only
subprocess_env (which is was gets turned into the output variables by
mod_env) is accessed by apr_table_set and apr_table_get ; these are
case-agnostic (the documentation for "key" says "case does not
matter"), but the case is preserved of the first time the variable is
set.

So we can set AAPRINCIPAL (and thus get the correct output case), and
get AAPrincipal elsewhere in the code.