no longer override Apache's default 401 page text
Apache lacks a proper core API for modules to override the default
text, so this was a fragile hack. Instead, advise users to use

  AuthzSendForbiddenOnFailure On

such that authorization failure is reported via the more appropriate
403 status code and page. This will make future breakage such as issue #11
less likely whenever Apache next changes their ErrorDocument data structures.

Note that AuthzSendForbiddenOnFailure was only introduced in Apache 2.3.11,
so Apache 2.2 users may prefer not to apply it.